+++ src/main/java/micdoodle8/mods/galacticraft/api/power/IEnergyStorageGC.java
@@ -9,7 +9,7 @@ public interface IEnergyStorageGC
 	 * @param simulate If true, the transfer will only be simulated.
 	 * @return The amount of energy that was successfully received (or would have been, if simulated).
 	 */
-	public int receiveEnergyGC(int amount, boolean simulate);
+	public float receiveEnergyGC(float amount, boolean simulate);
 
 	/**
 	 * Remove energy from the storage.
@@ -18,15 +18,15 @@ public interface IEnergyStorageGC
 	 * @param simulate If true, the transfer will only be simulated.
 	 * @return The amount of energy that was successfully extracted (or would have been, if simulated).
 	 */
-	public int extractEnergyGC(int amount, boolean simulate);
+	public float extractEnergyGC(float amount, boolean simulate);
 		
 	/**
 	 * Returns the amount of energy stored
 	 */
-	public int getEnergyStoredGC();
+	public float getEnergyStoredGC();
 	
 	/**
 	 * Returns the maximum amount of energy stored
 	 */
-	public int getCapacityGC();
+	public float getCapacityGC();
 }
